+++ src/main/java/cn/iwuliao/ds/service/HeloService.java
@@ -15,7 +15,7 @@ public interface HeloService {
 
     Integer hia();
 
-    Integer commit();
+    Long commit();
 
     Integer rollback();
